# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RiskIntel360 Platform API
  description: |
    RiskIntel360 is an autonomous multi-agent business intelligence platform that leverages Amazon Bedrock Nova 
    and AWS AI services to revolutionize enterprise market validation. The platform transforms traditional 
    manual market research (taking weeks) into intelligent, automated workflows that deliver comprehensive 
    business validation insights in under 2 hours.
    
    ## Key Features
    - **Multi-Agent Architecture**: Six specialized AI agents working in coordination
    - **Real-time Processing**: Complete validation workflows in under 2 hours
    - **Enterprise Security**: OAuth 2.0 authentication with role-based access control
    - **Scalable Infrastructure**: Auto-scaling from 3-50 instances based on demand
    
    ## Agent Types
    - **Market Research Agent**: Market data aggregation and trend analysis
    - **Competitive Intelligence Agent**: Competitor analysis and positioning assessment
    - **Financial Validation Agent**: Revenue projections and investment analysis
    - **Risk Assessment Agent**: Business risk evaluation and compliance analysis
    - **Customer Intelligence Agent**: Customer behavior analysis and market demand validation
    - **Synthesis & Recommendation Agent**: Cross-agent insight integration and strategic recommendations
  version: 1.0.0
  contact:
    name: RiskIntel360 Platform Team
    email: support@RiskIntel360.com
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - OAuth2: [read, write]
paths:
...
  /validations:
    post:
      summary: Create a new validation request
      description: |
        Initiates a new business validation workflow. The system will coordinate multiple AI agents
        to analyze market conditions, competitive landscape, financial viability, risks, and customer intelligence.
      tags:
        - Validations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRequest'
            examples:
              startup_validation:
                summary: Startup Market Validation
                value:
                  business_concept: "AI-powered personal finance management app"
                  target_market: "Young professionals aged 25-35 in North America"
                  analysis_scope: ["market", "competitive", "financial", "risk", "customer"]
                  priority: "high"
                  custom_parameters:
                    industry_focus: "fintech"
                    geographic_scope: "north_america"
                    timeline: "12_months"
      responses:
        '201':
          description: Validation request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List validation requests
      description: Retrieve a paginated list of validation requests for the authenticated user
      tags:
        - Validations
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filter by validation status
          schema:
            type: string
            enum: [pending, in_progress, completed, failed]
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            type: string
            enum: [low, medium, high, urgent]
      responses:
        '200':
          description: List of validation requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.RiskIntel360.com/oauth/authorize
          tokenUrl: https://auth.RiskIntel360.com/oauth/token
          scopes:
            read: Read access to validation data
            write: Create and modify validation requests
            admin: Administrative access
  schemas:
    ValidationRequest:
      type: object
      required:
        - business_concept
        - target_market
        - analysis_scope
      properties:
        business_concept:
          type: string
          description: Description of the business concept to validate
          example: "AI-powered personal finance management app"
          minLength: 10
          maxLength: 1000
        target_market:
          type: string
          description: Target market description
          example: "Young professionals aged 25-35 in North America"
          minLength: 10
          maxLength: 500
        analysis_scope:
          type: array
          description: Types of analysis to perform
          items:
            type: string
            enum: [market, competitive, financial, risk, customer]
          minItems: 1
          example: ["market", "competitive", "financial"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
          description: Priority level for the validation request
        deadline:
          type: string
          format: date-time
          description: Optional deadline for completion
        custom_parameters:
          type: object
          description: Additional parameters for customizing the analysis
          additionalProperties: true
          example:
            industry_focus: "fintech"
            geographic_scope: "north_america"
    ValidationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the validation request
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
          description: Current status of the validation
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time
        created_at:
          type: string
          format: date-time
          description: When the validation was created
        websocket_url:
          type: string
          format: uri
          description: WebSocket URL for real-time updates
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Unique request identifier for tracking

    # Additional schemas for completeness
...